feat: 프로필 수정 관련 type 정의 및 적용
